+++ AddOns/Altoholic/CHANGELOG.md
@@ -1,6 +1,7 @@
 # Altoholic Classic - Fork by Teelo
 
-## [Classic-v1.13.5-042](https://github.com/teelolws/Altoholic-Classic/tree/Classic-v1.13.5-042) (2020-09-06)
-[Full Changelog](https://github.com/teelolws/Altoholic-Classic/compare/Classic-v1.13.5-041...Classic-v1.13.5-042) [Previous Releases](https://github.com/teelolws/Altoholic-Classic/releases)
+## [Classic-v1.13.5-043](https://github.com/teelolws/Altoholic-Classic/tree/Classic-v1.13.5-043) (2020-09-16)
+[Full Changelog](https://github.com/teelolws/Altoholic-Classic/compare/Classic-v1.13.5-042...Classic-v1.13.5-043) [Previous Releases](https://github.com/teelolws/Altoholic-Classic/releases)
 
-- Fixed bug preventing raid lockouts from being stored properly.  
+- Classic-v1.13.5-043  
+- Summary - added ability to filter classes by their armor category (cloth/leather/etc).